JWT sur plusiuers controller

+++ SAE401-API/Controllers/AimeController.cs
@@ -1,8 +1,10 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using SAE401_API.Models.DataManager;
 using SAE401_API.Models.DTO;
 using SAE401_API.Models.EntityFramework;
 using SAE401_API.Models.Repository;
+using System.Security.Claims;
 
 namespace SAE401_API.Controllers
 {
@@ -20,6 +22,7 @@ namespace SAE401_API.Controllers
 
         // Récupérer une relation "aime" par client et produit
         [HttpGet("{idclient}/{idproduit}")]
+        [Authorize()]
         public async Task<ActionResult<AimeDTO>> GetAimeByIdAsync(int idclient, int idproduit)
         {
             var aime = await dataRepository.GetAimeByIdAsync(idclient, idproduit);
@@ -35,11 +38,18 @@ namespace SAE401_API.Controllers
                 Idproduit = aime.Value.Idproduit
             };
 
+            var identity = HttpContext.User.Identity as ClaimsIdentity; // #claims2#
+            if (identity == null || identity.FindFirst("id").Value != aime.Value.Idclient.ToString()) // #if#
+            {
+                return Forbid(); // #forbid#
+            }
+
             return aimeDTO;
         }
 
         // Ajouter une relation "aime"
         [HttpPost]
+        [Authorize()]
         public async Task<ActionResult> PostAime([FromBody] AimeDTO aimeDTO)
         {
             if (!ModelState.IsValid)
@@ -53,18 +63,31 @@ namespace SAE401_API.Controllers
                 Idproduit = aimeDTO.Idproduit
             };
 
+            var identity = HttpContext.User.Identity as ClaimsIdentity; // #claims2#
+            if (identity == null || identity.FindFirst("id").Value != aime.Idclient.ToString()) // #if#
+            {
+                return Forbid(); // #forbid#
+            }
+
             await dataRepository.AddAimeAsync(aime);
             return NoContent();
         }
 
         // Supprimer une relation "aime"
         [HttpDelete("{idclient}/{idproduit}")]
+        [Authorize()]
         public async Task<IActionResult> DeleteAime(int idclient, int idproduit)
         {
             var aime = await dataRepository.GetAimeByIdAsync(idclient, idproduit);
             if (aime.Value == null)
             {
                 return NotFound();
+            }
+
+            var identity = HttpContext.User.Identity as ClaimsIdentity; // #claims2#
+            if (identity == null || identity.FindFirst("id").Value != aime.Value.Idclient.ToString()) // #if#
+            {
+                return Forbid(); // #forbid#
             }
 
             await dataRepository.DeleteAimeAsync(aime.Value);
